+++ Source/CRUD.swift
@@ -1,7 +1,7 @@
 //
 // CRUD.swift
 //
-// Copyright (c) 2015 Recrea (http://recreahq.com/)
+// Copyright (c) 2015-2016 Recrea (http://recreahq.com/)
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -20,8 +20,6 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-
-import Alamofire
 
 
 // MARK: - CRUD
@@ -107,7 +105,7 @@ enum CRUDRequest<R: Resource>: Request {
 
   // MARK: Request
 
-  var method: Alamofire.Method {
+  var method: HTTPMethod {
     switch self {
     case .Create:
       return .POST