+++ Source/Item.swift
@@ -1,7 +1,7 @@
 //
 // Item.swift
 //
-// Copyright (c) 2015 Recrea (http://recreahq.com/)
+// Copyright (c) 2015-2016 Recrea (http://recreahq.com/)
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -34,7 +34,6 @@ public struct Item: Resource {
   public static let name = "items"
 
 }
-
 
 extension Item: CRUD {