+++ Tests/ResourceTests.swift
@@ -1,7 +1,7 @@
 //
 // ResourceTests.swift
 //
-// Copyright (c) 2015 Recrea (http://recreahq.com/)
+// Copyright (c) 2015-2016 Recrea (http://recreahq.com/)
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -24,7 +24,6 @@
 @testable import Quaderno
 
 import XCTest
-import Alamofire
 
 
 class ResourceTests: XCTestCase {